Decouple array position with number
We were really shoving extra data into the array index which required
the +1 / -1 operations to get to the index position.  Arrays are zero
based, but we started the numbers at 1, so we originally had to subtract
1 so that the start of the array represented the first item in the
index.  But if the min value was two, we'd start to run into trouble.
Then we'd try to push an item to the first position of an array that
does not yet have a value there yet.

Instead, the array is just a set of objects that each expose two
properties:
{ number : "", instances: ""}

Then when we get an array item, we can check the value for each.