Moved common function to separate file
Some of our business logic that we want to drive the application doens't
care if it's on page one or page two. We'd like to re-use as much as
possible.  This is why separating out DOM manipulation from our
functions was so important earlier.  That DOES make it care about which
page it's on.  Now that it's gone.  We can add these functions in a
separate file and include them before we start using them.

Step two, is incorporating into our other script file.