Removed Prompt
Prompt boxes have gone the way of the 8bit gif, blink, and marquee.
They are okay for test demos, but have no place in the modern web.
Users hate pop-ups and will try to ignore them and they poorly persist
relevant data.  Instead, just have a text box to store the value and
retrieve it.

In regards to copy, try not to use the word here for links and buttons.
It's not descriptive enough.  Of course it's here...it's a button.

See also: http://www.w3.org/QA/Tips/noClickHere